Move main into sub binary
